fix(math): explicitly cast fill value to uint8_t to avoid MSVC narrowing warning
【动机】
在分段筛法中，`std::fill(sieve.begin(), sieve.end(), 1);` 中 `1` 为 `int` 类型，而 `sieve` 是 `std::vector<uint8_t>`。MSVC 在 `/WX`（将警告视为错误）下将 `int` 转换为 `uint8_t` 视为可能丢失数据的警告，导致编译失败。

【原理】
将 `std::fill` 的第三个参数显式转换为 `uint8_t`：`static_cast<uint8_t>(1)`，消除隐式转换警告，使代码在 MSVC 严格警告级别下正常编译。

+++ src/math/BigInt.h
@@ -471,7 +471,7 @@ namespace jc {
             if (low % 2 == 0) low++;
 
             while (primes_found < count) {
-                std::fill(sieve.begin(), sieve.end(), 1);
+                std::fill(sieve.begin(), sieve.end(), static_cast<uint8_t>(1));
                 uint64_t high = low + S * 2 - 2;
                 
                 for (uint32_t p : base_primes) {
